feat(repo) move common/* to app/, rename app > client
