+++ .flowconfig
@@ -7,6 +7,5 @@
 
 [options]
 module.system.node.resolve_dirname=node_modules
-module.system.node.resolve_dirname=app
-module.system.node.resolve_dirname=common
-module.name_mapper.extension='scsshbs' -> '<PROJECT_ROOT>/common/flow/SCSSModuleStub.js'
+module.system.node.resolve_dirname=client
+module.name_mapper.extension='scsshbs' -> '<PROJECT_ROOT>/client/typedef/SCSSModuleStub.js'
