+++ client/components/spinkit/5.examples.md
