+++ client/components/spinkit/5.jsx
