+++ client/components/spinkit/index.js
