+++ client/typedef/FluxStandardAction.js
