+++ client/typedef/SCSSModuleStub.js
